# go-client/model_book.go
/*
 * Library API Spec
 *
 * This is an api for a library project that will be used in our SLO course.
 *
 * API version: v0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Book struct {
	BookID string `json:"bookID,omitempty"`
	BookDescription string `json:"bookDescription,omitempty"`
	BookTitle string `json:"bookTitle,omitempty"`
	BookAuthor string `json:"bookAuthor,omitempty"`
}