# go-server-server/go/model_delete_book_response.go
/*
 * Library API Spec
 *
 * This is an api for a library project that will be used in our SLO course.
 *
 * API version: v0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type DeleteBookResponse struct {

	Status string `json:"status,omitempty"`

	Message string `json:"message,omitempty"`
}